+++ src/main/java/com/handu/open/dubbo/monitor/domain/StatsTable.java
@@ -24,8 +24,8 @@ public class StatsTable implements Comparable<StatsTable> {
 
 	private long methodId;
 	private String serviceName, methodName;
-	private int successProvider, failureProvider, elapsedProvider;
-	private int successConsumer, failureConsumer, elapsedConsumer;
+	private long successProvider, failureProvider, elapsedProvider;
+	private long successConsumer, failureConsumer, elapsedConsumer;
 
 	public StatsTable(DayStats ds) {
 		this.successProvider = ds.getSuccessProvider();
@@ -37,6 +37,7 @@ public class StatsTable implements Comparable<StatsTable> {
 		this.methodId = ds.getMethodId();
 	}
 
+	
 	/**
 	 * @return the methodId
 	 */
@@ -44,13 +45,14 @@ public class StatsTable implements Comparable<StatsTable> {
 		return methodId;
 	}
 
+
 	/**
-	 * @param methodId
-	 *            the methodId to set
+	 * @param methodId the methodId to set
 	 */
 	public void setMethodId(long methodId) {
 		this.methodId = methodId;
 	}
+
 
 	/**
 	 * @return the serviceName
@@ -59,13 +61,14 @@ public class StatsTable implements Comparable<StatsTable> {
 		return serviceName;
 	}
 
+
 	/**
-	 * @param serviceName
-	 *            the serviceName to set
+	 * @param serviceName the serviceName to set
 	 */
 	public void setServiceName(String serviceName) {
 		this.serviceName = serviceName;
 	}
+
 
 	/**
 	 * @return the methodName
@@ -74,103 +77,110 @@ public class StatsTable implements Comparable<StatsTable> {
 		return methodName;
 	}
 
+
 	/**
-	 * @param methodName
-	 *            the methodName to set
+	 * @param methodName the methodName to set
 	 */
 	public void setMethodName(String methodName) {
 		this.methodName = methodName;
 	}
 
+
 	/**
 	 * @return the successProvider
 	 */
-	public int getSuccessProvider() {
+	public long getSuccessProvider() {
 		return successProvider;
 	}
 
+
 	/**
-	 * @param successProvider
-	 *            the successProvider to set
+	 * @param successProvider the successProvider to set
 	 */
-	public void setSuccessProvider(int successProvider) {
+	public void setSuccessProvider(long successProvider) {
 		this.successProvider = successProvider;
 	}
+
 
 	/**
 	 * @return the failureProvider
 	 */
-	public int getFailureProvider() {
+	public long getFailureProvider() {
 		return failureProvider;
 	}
 
+
 	/**
-	 * @param failureProvider
-	 *            the failureProvider to set
+	 * @param failureProvider the failureProvider to set
 	 */
-	public void setFailureProvider(int failureProvider) {
+	public void setFailureProvider(long failureProvider) {
 		this.failureProvider = failureProvider;
 	}
+
 
 	/**
 	 * @return the elapsedProvider
 	 */
-	public int getElapsedProvider() {
+	public long getElapsedProvider() {
 		return elapsedProvider;
 	}
 
+
 	/**
-	 * @param elapsedProvider
-	 *            the elapsedProvider to set
+	 * @param elapsedProvider the elapsedProvider to set
 	 */
-	public void setElapsedProvider(int elapsedProvider) {
+	public void setElapsedProvider(long elapsedProvider) {
 		this.elapsedProvider = elapsedProvider;
 	}
+
 
 	/**
 	 * @return the successConsumer
 	 */
-	public int getSuccessConsumer() {
+	public long getSuccessConsumer() {
 		return successConsumer;
 	}
 
+
 	/**
-	 * @param successConsumer
-	 *            the successConsumer to set
+	 * @param successConsumer the successConsumer to set
 	 */
-	public void setSuccessConsumer(int successConsumer) {
+	public void setSuccessConsumer(long successConsumer) {
 		this.successConsumer = successConsumer;
 	}
+
 
 	/**
 	 * @return the failureConsumer
 	 */
-	public int getFailureConsumer() {
+	public long getFailureConsumer() {
 		return failureConsumer;
 	}
 
+
 	/**
-	 * @param failureConsumer
-	 *            the failureConsumer to set
+	 * @param failureConsumer the failureConsumer to set
 	 */
-	public void setFailureConsumer(int failureConsumer) {
+	public void setFailureConsumer(long failureConsumer) {
 		this.failureConsumer = failureConsumer;
 	}
+
 
 	/**
 	 * @return the elapsedConsumer
 	 */
-	public int getElapsedConsumer() {
+	public long getElapsedConsumer() {
 		return elapsedConsumer;
 	}
 
+
 	/**
-	 * @param elapsedConsumer
-	 *            the elapsedConsumer to set
+	 * @param elapsedConsumer the elapsedConsumer to set
 	 */
-	public void setElapsedConsumer(int elapsedConsumer) {
+	public void setElapsedConsumer(long elapsedConsumer) {
 		this.elapsedConsumer = elapsedConsumer;
 	}
+
 
 	public void addDayStats(DayStats ds) {
 		this.successProvider += ds.getSuccessProvider();
